refctor: extract validation error
