+++ src/application/api/errors/ValidationError.ts
@@ -1,0 +1,11 @@
+export class ValidationError extends Error {
+  constructor(
+    public message: string
+  ){
+    super(message);
+  }
+
+  getMessage(){
+    return this.message;
+  }
+}
