+++ src/application/api/interfaces/index.ts
@@ -10,6 +10,6 @@ export interface HttpResponse<T = {}> {
 export interface BodyValidator<T> {
   validate(body: Record<string, any>): T
 }
-export interface Controller<T=any> {
-  handle(request: HttpRequest): Promise<HttpResponse<T>>
+export interface Controller<ResponseType=any> {
+  handle(request: HttpRequest): Promise<HttpResponse<ResponseType>>
 }
