+++ src/domain/usecases/CreateUserUseCase/CreateUserUseCase.ts
@@ -1,4 +1,4 @@
-import { ValidationError } from "@application/api/usecases/CreateUser/CreateUserBodyValidator";
+import { ValidationError } from "@application/api/errors/ValidationError";
 import { Encrypter } from "@domain/common/Encrypter";
 import { PortValidator } from "@domain/common/PortValidator";
 import { TokenGenerator } from "@domain/common/TokenGenerator";
